added global exception handling using RestAdvice. Added property rating to all record. Cleared  annotations in controller. Added all proper unit tests for all hardware services

+++ backend/src/main/java/de/mightypc/backend/model/specs/CPU.java
@@ -12,9 +12,9 @@ public record CPU(
         float price,
         int performance,
         int energyConsumption,
-        double rating
+        float rating
 ) {
-    public CPU(String name, String description, float price, int performance, int energyConsumption, double rating) {
+    public CPU(String name, String description, float price, int performance, int energyConsumption, float rating) {
         this(UUID.randomUUID().toString(), name, description, price, performance, energyConsumption, rating);
     }
 }
