+++ backend/src/main/java/de/mightypc/backend/model/specs/createspecs/CreateGpu.java
@@ -6,6 +6,6 @@ public record CreateGpu (
         float price,
         int performance,
         int energyConsumption,
-        double rating
+        float rating
 ){
 }
